add flexible scoring to general API
